+++ src/utils_gen/structures_utils.jl
@@ -1,7 +1,7 @@
 """
     function installed()
 
-Get a list of installed packages (from the depricated Pkg.installed())
+Get a list of installed packages (from the depricated Pkg.installed()).
 """
 function installed()
     deps = Pkg.dependencies()
@@ -283,7 +283,7 @@ end
     get_resource_colors(l::Vector{Transmission}, id_to_color_map::Dict{Any,Any})
 
 Get the colors linked to the resources in the transmission `l` (from modes(Transmission))
-based on the mapping `id_to_color_map`
+based on the mapping `id_to_color_map`.
 """
 function get_resource_colors(l::Transmission, id_to_color_map::Dict{Any,Any})
     resources::Vector{Resource} = [map_trans_resource(mode) for mode ∈ l.modes]